Move trainer to modules.trainer
